% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Check File Existence}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that gives the name of the file that the function will check and read. If it does not contain an absolute path, the file name is relative to the current working directory, getwd().}
}
\value{
This function returns a data frame that will be deprecated or an error message. As a side effect, the function will also show few rows of the data frame in the table format.
}
\description{
This is a simple function that check the existence of the file, and read the file.
}
\details{
The input has to be character within quotation marks, or it will lead to the error message. If the file does not exist in the current working dirctory, it will also lead to error message. If the file exists, it will read the file and show it in table format, and then deprecate it.
}
\examples{

\dontrun{
fars_read('accident_2013.csv.bz2')
}

}
