% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Years and Month}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
This function returns NULL or a list of data frames that have 2 columns, year and MONTH. The list has the same length as the input, years.
}
\description{
This is a function that select year and month columns from files of each year that is included in the input list.
}
\details{
If input list contains the year, the file of which does not exist, the corresponding output element of that year will be NULL, and the warning will be generated to show the invalid year.
}
\examples{

\dontrun{
fars_read_years(c(2011,2014,2017))
}

}
