% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create Summary Table of Year and Month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years.}
}
\value{
This function returns a table that shows the number of observation of each month in each year.
}
\description{
This is a function that generate a summary table of number of observations of
each year and each month.
}
\details{
If input list contains the year, the file of which does not exist, the warning will be generated to show the invalid year.
}
\examples{
\dontrun{
fars_summarize_years(c(2011,2014,2017))
}

}
